% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report}
\alias{generate_report}
\title{Generate a report from template}
\usage{
generate_report(report_template_name, output_file = NULL, param_list,
  see_now = FALSE)
}
\arguments{
\item{report_template_name}{name of template.}

\item{output_file}{name of the output file.}

\item{param_list}{A list of named parameters.}

\item{see_now}{Will launch browser if TRUE.}
}
\value{
invisible
}
\description{
Generate a report from template
}
\examples{
generate_report("report", output_file = "setosa-report.html", param_list = list(species = "setosa"))
}
